#145: Move devpack annotations to own package.
