+++ devpack/src/main/java/io/neow3j/devpack/compiler/Compiler.java
@@ -5,9 +5,9 @@ import static java.nio.charset.StandardCharsets.UTF_8;
 import io.neow3j.constants.InteropServiceCode;
 import io.neow3j.constants.OpCode;
 import io.neow3j.contract.ScriptBuilder;
-import io.neow3j.devpack.framework.EntryPoint;
-import io.neow3j.devpack.framework.Syscall;
-import io.neow3j.devpack.framework.Syscall.Syscalls;
+import io.neow3j.devpack.framework.annotations.EntryPoint;
+import io.neow3j.devpack.framework.annotations.Syscall;
+import io.neow3j.devpack.framework.annotations.Syscall.Syscalls;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
