+++ devpack/src/main/java/io/neow3j/devpack/framework/SmartContract.java
@@ -1,5 +1,7 @@
 package io.neow3j.devpack.framework;
 
+import io.neow3j.devpack.framework.annotations.Instruction;
+import io.neow3j.devpack.framework.annotations.Syscall;
 public abstract class SmartContract {
 
 }
