+++ devpack/src/main/java/io/neow3j/devpack/framework/annotations/EntryPoint.java
@@ -1,4 +1,4 @@
-package io.neow3j.devpack.framework;
+package io.neow3j.devpack.framework.annotations;
 
 import java.lang.annotation.*;
