+++ devpack/src/main/java/io/neow3j/devpack/framework/annotations/Manifest.java
@@ -1,4 +1,4 @@
-package io.neow3j.devpack.framework;
+package io.neow3j.devpack.framework.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Target;
