+++ devpack/src/main/java/io/neow3j/devpack/framework/annotations/Syscall.java
@@ -1,6 +1,6 @@
-package io.neow3j.devpack.framework;
+package io.neow3j.devpack.framework.annotations;
 
-import io.neow3j.devpack.framework.Syscall.Syscalls;
+import io.neow3j.devpack.framework.annotations.Syscall.Syscalls;
 import java.lang.annotation.Repeatable;
 
 @Repeatable(Syscalls.class)
